+++ actionview/test/actionpack/controller/render_test.rb
@@ -650,7 +650,7 @@ class TestController < ApplicationController
              "accessing_params_in_template_with_layout",
              "render_with_explicit_template",
              "render_with_explicit_string_template",
-             "update_page", "update_page_with_instance_variables"
+             "update_page", "update_page_with_instance_variables", "render_action_hello_world"
 
           "layouts/standard"
         when "action_talk_to_layout", "layout_overriding_layout"